Explicit require of cucumber-rails components
cucumber-rails entypoint tries to be too smart in my opinion by for
example trying to boot the app for you (which we are already doing
anyways and it's surprising), autodetecting the helper file that does
the require and fail if it's not called `env.rb` (which I have no idea
why), or throwing some warnings if certain settings are configured
(which in our case leads to an unexpected warning). Let's instead only
require the components we need.

This change fixes the long standing warning during class reloading
scenarios:

```
WARNING: You have set Rails' config.cache_classes to false (most
likely in config/environments/cucumber.rb). This setting is known to
cause problems with database transactions. Set config.cache_classes to
true if you want to use transactions.
```